+++ src-ui/app/views/dashboard-view/views/settings-integrations-view/settings-integrations-view.component.ts
@@ -5,6 +5,8 @@ import { hshrink } from 'src-ui/app/utils/animations';
 import { PulsoidService } from '../../../../services/integrations/pulsoid.service';
 import { VRChatService } from '../../../../services/vrchat.service';
 import { PULSOID_REFERRAL_ID } from 'src-ui/app/globals';
+import { ModalService } from '../../../../services/modal.service';
+import { MqttConfigModalComponent } from '../../../../components/mqtt-config-modal/mqtt-config-modal.component';
 
 @Component({
   selector: 'app-settings-integrations-view',
@@ -13,12 +15,15 @@ import { PULSOID_REFERRAL_ID } from 'src-ui/app/globals';
   animations: [hshrink()],
 })
 export class SettingsIntegrationsViewComponent {
-  deobfuscatePulsoidUsername = false;
   deobfuscated: string[] = [];
   deobfuscationTimers: { [service: string]: any } = {};
   copiedToClipboard: string[] = [];
 
-  constructor(protected pulsoid: PulsoidService, protected vrchat: VRChatService) {}
+  constructor(
+    protected pulsoid: PulsoidService,
+    protected vrchat: VRChatService,
+    private modalService: ModalService
+  ) {}
 
   protected deobfuscate(service: string) {
     if (!this.deobfuscated.includes(service)) this.deobfuscated.push(service);
@@ -48,5 +53,9 @@ export class SettingsIntegrationsViewComponent {
         warn('Tried copying link for unknown service');
         break;
     }
+  }
+
+  protected showMqttConfigModal() {
+    this.modalService.addModal(MqttConfigModalComponent).subscribe();
   }
 }