+++ src-ui/app/views/dashboard-view/views/settings-notifications-view/settings-notifications-view.component.ts
@@ -30,6 +30,10 @@ export class SettingsNotificationsViewComponent implements OnInit {
       label: 'XSOverlay',
     },
     {
+      id: 'OVRTOOLKIT',
+      label: 'OVRToolkit',
+    },
+    {
       id: 'DESKTOP',
       label: 'Windows',
     },